+++ app/apis/curriculums.py
@@ -7,10 +7,7 @@ from fastapi import APIRouter, Depends, HTTPException, Path, Query
 from sqlalchemy.orm import Session
 from starlette import status
 
-from schemas.curriculums import ReviewRequestListResponseBody, CurriculumDetailResponseBody, \
-                                QuestionCreateRequestBody, QuestionCreateResponseBody,\
-                                ReviewRequestCreateResponseBody, ReviewRequestCreateRequestBody, \
-                                QuestionListResponseBody
+from schemas.curriculums import CurriculumDetailResponseBody
 from cruds import curriculums as curriculums_crud
 
 logger = getLogger("uvicorn.app")
@@ -19,58 +16,6 @@ DbDependency = Annotated[Session, Depends(get_db)]
 
 router = APIRouter(prefix="/curriculums", tags=["Curriculums"])
 
-@router.get("/{curriculum_id}/reviews", response_model=ReviewRequestListResponseBody, status_code=status.HTTP_200_OK)
-async def find_review_list(db: DbDependency, curriculum_id: int = Path(gt=0)):
-    """
-    カリキュラムのレビュー一覧
-
-    Parameter
-    -----------------------
-    curriculum_id: int
-        レビュー一覧を取得したいカリキュラムのID
-
-    Returns
-    -----------------------
-    reviews: array
-        id: int
-            レビューリクエストのID
-        curriculum_id: int
-            関連するカリキュラムのID
-        user_id: int
-            レビューリクエストを投稿したユーザーのID
-        title: str
-            レビューリクエストのタイトル
-        content: str
-            レビューリクエストの内容
-        is_closed: bool
-            レビューリクエストがクローズされているかどうか
-        created_at: str
-            レビューリクエストが作成された日時（ISO 8601形式）
-        updated_at: str
-            レビューリクエストが最後に更新された日時（ISO 8601形式）
-    """
-
-    reviews = curriculums_crud.find_reviews_by_curriculum_id(db, curriculum_id)
-
-    li = []
-    for review in reviews:
-        di = {
-            "id": review.id,
-            "curriculum_id": review.curriculum_id,
-            "user_id": review.user_id,
-            "title": review.title,
-            "content": review.content,
-            "is_closed": review.is_closed,
-            "created_at": review.created_at.isoformat(),
-            "updated_at": review.updated_at.isoformat()
-        }
-        li.append(di)
-
-    re_di = {
-        "reviews": li
-    }
-
-    return re_di
 
 @router.get("/{curriculum_id}", response_model=CurriculumDetailResponseBody, status_code=status.HTTP_200_OK)
 async def find_curriculum_details(db: DbDependency, curriculum_id: int = Path(gt=0)):
@@ -159,239 +104,3 @@ async def find_curriculum_details(db: DbDependency, curriculum_id: int = Path(gt
 #         "tests": li
 #     }
 #     return re_di
-
-@router.post("/{course_id}/questions", response_model=QuestionCreateResponseBody, status_code=status.HTTP_201_CREATED)
-async def create_question(db: DbDependency, param: QuestionCreateRequestBody, course_id: int = Path(gt=0)):
-    """
-    質問投稿作成取得
-    
-    Parameter
-    -----------------------
-    course_id: int
-        質問を投稿したいコースのID
-    dict
-        curriculum_id: int
-            質問が紐づくカリキュラムのID
-        user_id: int
-            ユーザーのID
-        title: str
-            質問のタイトル
-        objective: str
-            学習内容で実践したこと
-        current_situation: str
-            現状
-        research: str
-            自分が調べたこと
-        content: str 
-            質問の内容
-        media_content: str
-            関連するメディアコンテンツの情報
-            url: str
-                メディアコンテンツのURL
-
-    Returns
-    -----------------------
-    dict
-        question_id: int
-            質問のID
-        curriculum_id: int
-            カリキュラムのID
-        user: dict
-            user_id: int
-                質問を投稿したユーザーのID
-            name: str
-                質問を投稿したユーザーの名前
-        title: str
-            質問のタイトル
-        objective: str
-            学習内容で実践したこと
-        current_situation: str
-            現状
-        research: str
-            自分が調べたこと
-        content: str 
-            質問の内容
-        media_content: dict
-            関連するメディアコンテンツの情報
-        created_at: str
-            質問作成日（ISO 8601形式）
-        is_read: bool
-            未読コメントの有無
-        is_closed: bool
-            完了しているかどうか
-        reply_counts: int
-            質問の返信数
-    """
-    found_curriculum = curriculums_crud.find_curriculum_by_curriculum_id(db, param.curriculum_id)
-
-    if not found_curriculum:
-        raise HTTPException(status_code=404, detail="Curriculum not found.")
-
-    li = []
-    datas = param.media_content
-    for data in datas:
-        if hasattr(data, "url"):
-            di = {
-                "url": data.url
-            }
-            li.append(di)
-
-    media_json = li
-
-    try:
-        new_question = curriculums_crud.create_question(db, course_id, param, media_json)
-        db.commit()
-
-        user = curriculums_crud.find_user_by_id(db, new_question.user_id)
-        
-        re_di = {
-            "question_id": new_question.id,
-            "curriculum_id": new_question.curriculum_id,
-            "user": {
-                "user_id": user.id,
-                "name": user.last_name + user.first_name
-            },
-            "title": new_question.title,
-            "objective": new_question.objective,
-            "current_situation": new_question.current_situation,
-            "research": new_question.research,
-            "content": new_question.content,
-            "media_content": new_question.media_content,
-            "created_at": new_question.created_at.isoformat(),
-            "is_read": False, # 作成なので、この時点ではFalse確定
-            "is_closed": False, # 作成なので、この時点ではFalse確定
-            "reply_counts": 0 # 作成なので、この時点では0確定
-        }
-
-        return re_di
-
-    except Exception as e:
-        logger.error(str(e)) 
-        db.rollback()
-        raise HTTPException(status_code=400, detail="Invalid input data.")
-
-@router.get("/{curriculum_id}/questions", response_model=QuestionListResponseBody, status_code=status.HTTP_200_OK)
-async def find_question_list_in_curriculum(db: DbDependency, curriculum_id: int):
-    """
-    カリキュラムの質問一覧
-    
-    Parameter
-    -----------------------
-    curriculum_id: int
-        質問一覧を取得したいカリキュラムのID
-
-    Returns
-    -----------------------
-    dict
-        question_id: int
-            質問のID
-        curriculum_id: int
-            カリキュラムのID
-        user_id: int
-            ユーザーのID
-        title: str
-            質問のタイトル
-        content: str 
-            質問の内容
-        media_content: str
-            関連するメディアコンテンツの情報
-            url: str
-                メディアコンテンツのURL
-    """
-
-    questions = curriculums_crud.find_questions_by_curriculum_id(db, curriculum_id)
-
-    if not questions:
-        raise HTTPException(status_code=404, detail="Questions not found for the specified curriculum.")
-
-    li = []
-    for question in questions:
-        media_content_list = []
-
-        datas = question.media_content
-        
-        for data in datas:
-            if "url" in data:
-                di = {
-                    "url": data.get("url", "")
-                }
-                media_content_list.append(di)
-        
-        di = {
-            "question_id": question.id,
-            "curriculum_id": question.curriculum_id,
-            "user_id": question.user_id,
-            "title": question.title,
-            "content": question.content,
-            "media_content": media_content_list
-        }
-        li.append(di)
-    
-    re_di = {
-        "questions": li
-    }
-    return re_di
-
-
-@router.post("/{curriculum_id}/reviews", response_model=ReviewRequestCreateResponseBody, status_code=status.HTTP_201_CREATED)
-async def create_review_request(db: DbDependency, param: ReviewRequestCreateRequestBody, curriculum_id: int):
-
-    """
-    レビュー作成
-    
-    Parameter
-    -----------------------
-    curriculum_id: int
-        詳細を取得したいカリキュラムのID
-    dict
-        user_id: int
-            ユーザーのID
-        title: str
-            レビューリクエストのタイトル
-        content: str 
-            レビューリクエストの内容
-        is_closed: boolean
-            レビューリクエストの初期状態（通常はfalseで未クローズ状態）
-     Returns
-    -----------------------
-    dict
-        id: int
-            レビューリクエストのID
-        curriculum_id: int
-            カリキュラムのID
-        user_id: int
-            ユーザーのID
-        title: str
-            レビューリクエストのタイトル
-        content: str 
-            レビューリクエストの内容
-        is_closed: boolean
-            レビューリクエストがクローズされているかどうか（boolean）
-        created_at: str
-            作成された日時
-    """
-
-
-    found_curriculum = curriculums_crud.find_review_request_by_curriculum_id(db, curriculum_id)
-
-    if not found_curriculum:
-        raise HTTPException(status_code=404, detail="Curriculum not found.")
-
-    try:
-        reviews = curriculums_crud.create_review_request(db, curriculum_id, param.user_id, param.title, param.content, param.is_closed)
-        db.commit()
-        di = {
-            "id": reviews.id,
-            "curriculum_id": reviews.curriculum_id,
-            "user_id": reviews.user_id,
-            "title": reviews.title,
-            "content": reviews.content,
-            "is_closed": reviews.is_closed,
-            "created_at": reviews.created_at.isoformat()
-        }
-        
-        return di
-    except Exception as e:
-        logger.error(str(e)) 
-        db.rollback()
-        raise HTTPException(status_code=400, detail="Invalid input data.")
